# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Authentication and Authorization Micro-service for the FamTrust Web App.",
        "title": "FamTrust API Backend - Auth",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
        "/create-user": {
            "post": {
                "description": "Create a Sub-User/Member User Account",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-Accounts"
                ],
                "summary": "Create a Sub-User/Member User Account",
                "operationId": "create-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email of the new user",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password of the new user",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional Role ID string for new user. Defaults to 'member' if not specified",
                        "name": "roleID",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Optional true or false value to set new user 2FA preference",
                        "name": "has2FA",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError500"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Login to FamTrust (Supports 2FA by Email)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-Authentication"
                ],
                "summary": "Login to FamTrust (Supports 2FA by Email)",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "User Credentials",
                        "name": "Credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.loginRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "User 2FA Code",
                        "name": "2FACode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponse200"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError401"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError500"
                        }
                    }
                }
            }
        },
...
        "/signup": {
            "post": {
                "description": "Create an Admin/Main User Account",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-Accounts"
                ],
                "summary": "Create an Admin/Main User Account",
                "operationId": "signup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email of the new user",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password of the new user",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional true or false value to set new user 2FA preference",
                        "name": "has2FA",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError500"
                        }
                    }
                }
            }
        },
        "/users/by/default-group": {
            "get": {
                "description": "Get Users by Group ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Retrieve-Users"
                ],
                "summary": "Get Users by Group ID",
                "operationId": "users-by-groupID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID to filter Users By",
                        "name": "groupID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError500"
                        }
                    }
                }
            }
        },
        "/validate": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validate User Login Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-Authentication"
                ],
                "summary": "Validate User Login Token",
                "operationId": "validate",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.validateSampleResponse200"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError401"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handlers.loginSampleResponse200": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User Logged in successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                },
                "token": {
                    "type": "string",
                    "example": "b6d4a7e1d2d841a1afe874a2a5c15d8b"
                }
            }
        },
        "handlers.loginSampleResponseError401": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid Credentials"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 401
                }
            }
        },
        "handlers.loginSampleResponseError500": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "An error occured"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
...
        "handlers.validateSampleResponse200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "user": {
                            "$ref": "#/definitions/handlers.validateSampleResponse200User"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Request successful"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                },
                "token": {
                    "type": "string",
                    "example": "b6d4a7e1d2d841a1afe874a2a5c15d8b"
                }
            }
        },
        "handlers.validateSampleResponse200User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "has2FA": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "string",
                    "example": "d38f91b2-dc3b-4f9d-aeb4-7b95c91e9d08"
                },
                "isFreezed": {
                    "type": "boolean",
                    "example": true
                },
                "isVerified": {
                    "type": "boolean",
                    "example": true
                },
                "lastLogin": {
                    "type": "string",
                    "example": "2024-07-22T14:30:00Z"
                },
                "role": {
                    "$ref": "#/definitions/handlers.validateSampleResponseRole"
                }
            }
        },
        "handlers.validateSampleResponseRole": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "admin"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "canTransact",
                        " canWithdraw"
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}